+++ docs/changes.md
@@ -3,6 +3,14 @@ title: Model Change Log
 
 ---
 
+## Version 2.2 {#v2-2}
+
+### v2.2.0 {#v2-2-0}
+
+* Bump minimum version to PHP v5.4.0.
+* Add `Orchestra\Model\User::notify()` method which utilize `Orchestra\Notifier\NotifiableTrait`.
+* Move duplicates code to utilize traits.
+
 ## Version 2.1 {#v2-1}
 
 ### v2.1.5 {#v2-1-5}